Image service implemented to use image CDN + config folder moved under core
