+++ TravPhotBlog/ClientApp/src/app/core/configs/constants.ts
@@ -1,4 +1,5 @@
 export class Constants {
   public static get API_ROOT_URL(): string { return "http://travphotblog.southeastasia.cloudapp.azure.com/api"; }
   public static get SITE_TITLE(): string { return "Travel, Photography, Blogging"; }
+  public static get CLOUDIMG_BASEURL(): string { return "https://aukkymggen.cloudimg.io/v7/_blob_/"; }
 }
