+++ TravPhotBlog/ClientApp/src/app/core/services/image.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ImageService } from './image.service';
+
+describe('ImageService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: ImageService = TestBed.get(ImageService);
+    expect(service).toBeTruthy();
+  });
+});
