+++ TravPhotBlog/ClientApp/src/app/core/services/social.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Constants } from '../../configs/constants';
+import { Constants } from '../../core/configs/constants';
 import { InstaModel } from '../../shared/models/instapost';
 
 @Injectable({
