+++ TravPhotBlog/ClientApp/src/app/home/components/recent-posts/recent-posts.component.html
@@ -3,7 +3,7 @@
   <div class="col col-6 push-center">
     <div class="sayit_widget_title">
       <div class="sayit_widget_container sayit_text_align_center sayit_mb_60">
-        <h2 class="sayit_element_heading sayit_view_type_1">Enjoy of My Latest Travels</h2>
+        <h2 class="sayit_element_heading sayit_view_type_1">Latest Blogs</h2>
         <p class="sayit_subtitle_cont">Some of my latest travel stories.</p>
       </div>
     </div>
