+++ TravPhotBlog/ClientApp/src/app/home/home.component.html
@@ -51,7 +51,7 @@
         <!-- Socials -->
         <div class="row gutters">
           <div class="col col-6">
-            <img src="assets/imgs/Blogger/Priyabrata.jpg" alt="" class="sayit_md_mb_15" />
+            <img src="{{bloggerImageUrl}}" alt="Blogger image" class="sayit_md_mb_15" />
           </div>
 
           <div class="col col-6 sayit_pl_20 sayit_md_padding_0">
